Bunch More Work on HLSL parser.
This is not nearly ready for primetime. I doubt it does much more than
compile (and maybe only on my machine).

This reworks/improves/cleans up the grammar quite a bit, and fills in much of
the C code required to generate an Abstract Syntax Tree.

Nothing is done with this tree yet, including freeing it.